Upgraded to gradle 5 and spring boot 2
Did this to fix the fact that I was unable to get both kotlin and java files into the same jar.

Ended up not being able to solve that after upgrading either, and just converted the Pojo. So probably didn't need to do the upgrades. But was fun, and now everything is updated!