Added support for horizontal dragscrolling
Had a bit of a hard time figuring out how to prevent the scroll section growing outside its parent. Turned out the issue was related to the default flex box sizing options, which prevents flex-boxes from becoming smaller than it's contents. This is weirdly (?) fixed by setting min-width of the parent to 0.