Upgraded all the (backend) things!
Most notably, it was a bit of hassle getting airtable.java to still work, as it hasn't been updated in ages. It still deploys to jcentral, which is deprecated. Fortunately it will probably be hosted as read only for the indeterminable future, so I'll leave it at that for now.